requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.7
djangorestframework==3.12.2
dlib==19.21.99
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.20.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
